Move subscribe response types to actions directory
Relocated Zuora subscribe response schema and type from zuoraSchemas.ts to types/actions/subscribe.ts for better organization and separation of concerns.
